Fix a few warnings

+++ src-core/modules/aqua/deframer.cpp
@@ -83,7 +83,7 @@ namespace aqua
         return state;
     }
 
-    std::vector<std::array<uint8_t, CADU_SIZE>> CADUDeframer::work(uint8_t *input, size_t size)
+    std::vector<std::array<uint8_t, CADU_SIZE>> CADUDeframer::work(uint8_t *input, int size)
     {
         // Output buffer
         std::vector<std::array<uint8_t, CADU_SIZE>> frames;
